Grape based "helper" method + test
NORMAL rack `requests.cookies` is a hash, Grape's is a
`Grape::Cookies` type, which doesn't have the same API as a Ruby
Hash. As such, this library uses the methods that `Grape::Cookies`
actually emits/exposes/exports.

Since this code is used as a Mixin, it requires the code it's mixed
into contains `request`, `request.cookies`, etc. in the same way that
this code requires it, so we rely on the caller to use the right one.

The actual API "surface" is the same as the SnowplowRubyDuid::Helper.
By design, no attempt to "share code" with the original
SnowplowRubyDuid::Helper has been made.