"fixed" tests
This fixes an issue of "which test runs first wins" issue by not
re-using the same name for a helper class to mix-in the code into.
The helper mock class with the same name but different implementation
was causing whichever test than ran after the first one to use the
first one's helper class, which caused subsequent tests to fail.